refactor: 옛 /wish 페이지 잔재 정리 및 archive 콜로케이션 재배치
