+++ apps/web/src/app/archive/wish/_apis/deleteWishes.ts
