+++ apps/web/src/app/archive/wish/_components/WishCardSkeleton.tsx
