+++ apps/web/src/app/archive/wish/_components/WishlistFabArea.tsx
