+++ apps/web/src/app/tournament/[id]/create/by-wish/_components/WishSelectCard.tsx
@@ -1,6 +1,6 @@
-import WishCard from '@/app/archive/_components/wish-grid/WishCard';
 import CheckboxEmptyIconFill from '@/assets/icons/fill/checkbox-empty.svg';
 import CheckboxSelectedIconFill from '@/assets/icons/fill/checkbox-selected.svg';
+import WishCard from '@/components/common/wish-card';
 
 type WishSelectCardProps = {
   name: string;
